Updated: an artificers ascent.tas (6:47.592) from megamaz
https://discord.com/channels/1269042133541585049/1269042134124855321/1270993846918053949
An Artificer's Ascent draft in 6:47.592(23976)
I feel like there's a lot of huge skips that could still be done that I completely missed, and a lot of the elytra movement isn't optimized (see video, it's pretty obvious)
however right now I am tired, lmk if you plan on refining it otherwise I'll be working on it some more tomorrow 
A lot of the angles look funky, and that's because I had to convert them from AnalogMode Precise to AnalogMode Ignore using python. Also I lost 2 frames in the process of resyncing this error. 
https://www.youtube.com/watch?v=OxQEO_mKlXE